Fix KDBX4 key calculation on 32 bit platforms (#45)
* Fix #41: Fix incorrect key calculation on 32 bit platforms

* add 32bit test to ci

+++ src/parse/kdbx4.rs
@@ -273,7 +273,7 @@ pub(crate) fn decrypt_xml(
 
     // verify credentials
     let hmac_key = crypt::calculate_sha512(&[&header.master_seed, &transformed_key, b"\x01"])?;
-    let header_hmac_key = hmac_block_stream::get_hmac_block_key(usize::max_value(), &hmac_key)?;
+    let header_hmac_key = hmac_block_stream::get_hmac_block_key(u64::max_value(), &hmac_key)?;
     if header_hmac != crypt::calculate_hmac(&[header_data], &header_hmac_key)?.as_slice() {
         return Err(Error::IncorrectKey);
     }
